GetAvailablityByFeaturesSp
Repeated in both Additions and Build (I really should get rid of the
dates on these . . .).    Returns # of rooms available--this is rooms
that have not been checked out/in maintenance or have a reservation for
a specific room; RoomsReserved which is the number of reservations for
this type of room.  The reservations are right now working on the lowest
rate/ranked room basis.  Also returns if smoking is allowed, the rate,
beds (I hope to come back and clean this up into Bed type/Qty) and
Features, and an example room.