CC_Info
Still needs a view for the Address, but being that I tied to the
existing address, we will have to create some logic i.e. if same as
customer's bill-to, you would have a check box that would then just have
you pass 0 for the customer sequence,  otherwise it would create a new
sequence.