GetAvailableRoomsSp
This has been fixed so that it works for all three types.  There was a
reason that I avoided it before . . . . it has to do with room
duplicates.  There are 6-7 views that are new or changed in this + the
sp.  Update should load all that you need for the changes